Removed backwards ima18 compatibility
